Downloaded public updates from server and began account set API's

+++ Backend/src/main/java/com/troy/dramaserver/database/PointEntry.java
@@ -99,7 +99,7 @@ public class PointEntry implements Serializable {
 			if (rate == 0.0)
 				return amount;
 			else
-				return rate * (double) amount;
+				return rate * amount;
 		}
 
 		@Override
